# Make test_(batch_)beam_search.py faster.