@ShigekiKarita 2020-09-12
According to https://circleci.com/api/v1.1/project/github/espnet/espnet/21011/output/105/0?file=true&allocation-id=5f5c93195321443c7a225a2f-0-build%2F1192F48E
these tests got much faster in circle ci
- 35 ('test/test_beam_search.py', 0.6900000000000001)
- 39 ('test/test_batch_beam_search.py', 0.6000000000000002)

> master (#2459)
>  0 ('test/test_beam_search.py', 536.879999999997)
> 1 ('test/test_batch_beam_search.py', 289.2799999999999)