@ShigekiKarita 2020-09-12
> It is better to add TODO comments about this case.

I see. I added TODO and warning.